% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{ds_distribution}
\alias{ds_distribution}
\title{Get a distribution}
\usage{
ds_distribution(name = "stable", registry = ds_registry())
}
\arguments{
\item{name}{Distribution name (e.g. `"2026.04"`) or alias (`"stable"`,
`"testing"`).}

\item{registry}{A registry, see [ds_registry()].}
}
\value{
The distribution, an object of class `ds_distribution`.
}
\description{
Get a distribution
}
\examples{
ds_distribution("stable")
}
